Revert "String optimization and addition of Suffix() (#126)"
This reverts commit 7cb7b9a6faf6fba82441c45a81af58e8a2df0e90.
