+++ pyteal/ast/__init__.py
@@ -88,8 +88,7 @@ from .binaryexpr import (
 )
 
 # ternary ops
-from .ternaryexpr import Ed25519Verify, SetBit, SetByte
-from .substring import Substring, Extract, Suffix
+from .ternaryexpr import Ed25519Verify, Substring, Extract, SetBit, SetByte
 
 # more ops
 from .naryexpr import NaryExpr, And, Or, Concat
@@ -190,7 +189,6 @@ __all__ = [
     "Ed25519Verify",
     "Substring",
     "Extract",
-    "Suffix",
     "SetBit",
     "SetByte",
     "NaryExpr",
