+++ pyteal/ast/ternaryexpr_test.py
@@ -43,6 +43,73 @@ def test_ed25519verify_invalid():
         Ed25519Verify(Bytes("data"), Bytes("sig"), Int(0))
 
 
+def test_substring():
+    args = [Bytes("my string"), Int(0), Int(2)]
+    expr = Substring(args[0], args[1], args[2])
+    assert expr.type_of() == TealType.bytes
+
+    expected = TealSimpleBlock(
+        [
+            TealOp(args[0], Op.byte, '"my string"'),
+            TealOp(args[1], Op.int, 0),
+            TealOp(args[2], Op.int, 2),
+            TealOp(expr, Op.substring3),
+        ]
+    )
+
+    actual, _ = expr.__teal__(teal2Options)
+    actual.addIncoming()
+    actual = TealBlock.NormalizeBlocks(actual)
+
+    assert actual == expected
+
+
+def test_substring_invalid():
+    with pytest.raises(TealTypeError):
+        Substring(Int(0), Int(0), Int(2))
+
+    with pytest.raises(TealTypeError):
+        Substring(Bytes("my string"), Txn.sender(), Int(2))
+
+    with pytest.raises(TealTypeError):
+        Substring(Bytes("my string"), Int(0), Txn.sender())
+
+
+def test_extract():
+    args = [Bytes("my string"), Int(0), Int(2)]
+    expr = Extract(args[0], args[1], args[2])
+    assert expr.type_of() == TealType.bytes
+
+    expected = TealSimpleBlock(
+        [
+            TealOp(args[0], Op.byte, '"my string"'),
+            TealOp(args[1], Op.int, 0),
+            TealOp(args[2], Op.int, 2),
+            TealOp(expr, Op.extract3),
+        ]
+    )
+
+    actual, _ = expr.__teal__(teal5Options)
+    actual.addIncoming()
+    actual = TealBlock.NormalizeBlocks(actual)
+
+    assert actual == expected
+
+    with pytest.raises(TealInputError):
+        expr.__teal__(teal4Options)
+
+
+def test_extract_invalid():
+    with pytest.raises(TealTypeError):
+        Extract(Int(0), Int(0), Int(2))
+
+    with pytest.raises(TealTypeError):
+        Extract(Bytes("my string"), Txn.sender(), Int(2))
+
+    with pytest.raises(TealTypeError):
+        Extract(Bytes("my string"), Int(0), Txn.sender())
+
+
 def test_set_bit_int():
     args = [Int(0), Int(2), Int(1)]
     expr = SetBit(args[0], args[1], args[2])
